Integrate the Ready Mode prompt with IE and Chrome Frame. In Ready Mode, prompts are displayed when the user browses to GCF-enabled sites, allowing the user to permanently activate, permanently decline, or temporarily decline Chrome Frame.
BUG=None
TEST=chrome_frame_unittests --gtest_filter=Ready* && chrome_frame_unittests --gtest_filter=Infobar* && setup.exe --chrome-frame --ready-mode --multi-install --system--level --chrome
Review URL: http://codereview.chromium.org/6040003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@71215 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6e58a95b99e1c88172d5eae8925c8dfa39c2b765

+++ chrome/installer/util/install_util.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2010 The Chromium Authors. All rights reserved.
+// Copyright (c) 2011 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 //
@@ -116,15 +116,15 @@ void InstallUtil::WriteInstallerResult(bool system_install,
                                        int string_resource_id,
                                        const std::wstring* const launch_cmd) {
   const HKEY root = system_install ? HKEY_LOCAL_MACHINE : HKEY_CURRENT_USER;
-  int installer_result = (GetInstallReturnCode(status) == 0) ? 0 : 1;
+  DWORD installer_result = (GetInstallReturnCode(status) == 0) ? 0 : 1;
   scoped_ptr<WorkItemList> install_list(WorkItem::CreateWorkItemList());
   install_list->AddCreateRegKeyWorkItem(root, state_key);
   install_list->AddSetRegValueWorkItem(root, state_key,
                                        installer::kInstallerResult,
                                        installer_result, true);
   install_list->AddSetRegValueWorkItem(root, state_key,
-                                       installer::kInstallerError, status,
-                                       true);
+                                       installer::kInstallerError,
+                                       static_cast<DWORD>(status), true);
   if (string_resource_id != 0) {
     std::wstring msg = installer::GetLocalizedString(string_resource_id);
     install_list->AddSetRegValueWorkItem(root, state_key,
